+++ package.json
@@ -4,7 +4,8 @@
   "main": "app.js",
   "scripts": {
     "start": "NODE_ENV=production nodejs app.js",
-    "test": "PORT=3000 NODE_ENV=test nodejs app.js"
+    "test": "PORT=3000 NODE_ENV=test nodejs app.js",
+    "build-front": "webpack --config webpack.config.js"
   },
   "keywords": [
     "waterrower",
@@ -34,6 +35,8 @@
     "xmlbuilder": "^10.0.0"
   },
   "devDependencies": {
-    "pi-gpio": "0.0.8"
+    "pi-gpio": "0.0.8",
+    "webpack": "^4.17.1",
+    "webpack-cli": "^3.1.0"
   }
 }
