+++ public/index.html
@@ -63,6 +63,6 @@
 <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDjxms3dDPLsebcsMXg0STuDDdnqlPF6wA"
         async defer></script>
 <!-- Latest compiled and minified JavaScript -->
-<script src="/js/main.js"></script>
+<script src="/js/bundle.js"></script>
 </body>
 </html>
